Restructure universal wallet crates to be better organised (in a non-breaking way)
