+++ crates/universal-wallet/schema/src/lib.rs
@@ -1,8 +1,13 @@
-pub mod display;
+mod visitors;
+pub use visitors::display;
 #[cfg(feature = "serde")]
-pub mod json_to_borsh;
+pub use visitors::json_to_borsh;
+
 pub mod schema;
 pub mod ty;
+
+#[cfg(test)]
+mod tests;
 
 pub extern crate bech32;
