+++ crates/universal-wallet/schema/src/tests/schema.rs
