+++ crates/universal-wallet/schema/src/visitors/display.rs
