+++ crates/universal-wallet/schema/src/visitors/mod.rs
@@ -1,0 +1,3 @@
+pub mod display;
+#[cfg(feature = "serde")]
+pub mod json_to_borsh;
